converted math types to new pattern
