+++ src/math/aabb/line-intersection.test.ts
@@ -22,11 +22,10 @@ SOFTWARE.*/
 
 
 import { describe, it, expect } from 'vitest';
-import { lineIntersection } from './line-intersection.js';
-import { Aabb } from './aabb.js';
-import { Line3 } from '../line3/line3.js';
+import { Aabb } from '../index.js';
+import { Line3 } from '../index.js';
 
-describe('lineIntersection', () => {
+describe('Aabb.lineIntersection', () => {
     const box: Aabb = {
         min: [0, 0, 0],
         max: [2, 2, 2]
@@ -37,7 +36,7 @@ describe('lineIntersection', () => {
             a: [1, 1, 1],
             b: [3, 3, 3]
         };
-        expect(lineIntersection(box, line)).toBe(0);
+        expect(Aabb.lineIntersection(box, line)).toBe(0);
     });
 
     it('should return correct alpha when line intersects box from outside', () => {
@@ -45,7 +44,7 @@ describe('lineIntersection', () => {
             a: [-1, 1, 1],
             b: [3, 1, 1]
         };
-        expect(lineIntersection(box, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(box, line)).toBe(0.25);
     });
 
     it('should return -1 when line does not intersect box', () => {
@@ -53,7 +52,7 @@ describe('lineIntersection', () => {
             a: [3, 3, 3],
             b: [5, 5, 5]
         };
-        expect(lineIntersection(box, line)).toBe(-1);
+        expect(Aabb.lineIntersection(box, line)).toBe(-1);
     });
 
     it('should return -1 when line intersects but segment is too short', () => {
@@ -61,7 +60,7 @@ describe('lineIntersection', () => {
             a: [-2, 1, 1],
             b: [-1, 1, 1]
         };
-        expect(lineIntersection(box, line)).toBe(-1);
+        expect(Aabb.lineIntersection(box, line)).toBe(-1);
     });
 
     it('should handle line parallel to x-axis', () => {
@@ -69,7 +68,7 @@ describe('lineIntersection', () => {
             a: [-1, 1, 1],
             b: [3, 1, 1]
         };
-        expect(lineIntersection(box, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(box, line)).toBe(0.25);
     });
 
     it('should handle line parallel to y-axis', () => {
@@ -77,7 +76,7 @@ describe('lineIntersection', () => {
             a: [1, -1, 1],
             b: [1, 3, 1]
         };
-        expect(lineIntersection(box, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(box, line)).toBe(0.25);
     });
 
     it('should handle line parallel to z-axis', () => {
@@ -85,7 +84,7 @@ describe('lineIntersection', () => {
             a: [1, 1, -1],
             b: [1, 1, 3]
         };
-        expect(lineIntersection(box, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(box, line)).toBe(0.25);
     });
 
     it('should handle line that goes through corner', () => {
@@ -93,7 +92,7 @@ describe('lineIntersection', () => {
             a: [-1, -1, -1],
             b: [3, 3, 3]
         };
-        expect(lineIntersection(box, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(box, line)).toBe(0.25);
     });
 
     it('should handle line that starts at box boundary', () => {
@@ -101,7 +100,7 @@ describe('lineIntersection', () => {
             a: [0, 1, 1],
             b: [3, 1, 1]
         };
-        expect(lineIntersection(box, line)).toBe(0);
+        expect(Aabb.lineIntersection(box, line)).toBe(0);
     });
 
     it('should handle line that ends at box boundary', () => {
@@ -109,7 +108,7 @@ describe('lineIntersection', () => {
             a: [-1, 1, 1],
             b: [2, 1, 1]
         };
-        expect(lineIntersection(box, line)).toBe(1/3);
+        expect(Aabb.lineIntersection(box, line)).toBe(1 / 3);
     });
 
     it('should handle negative coordinates', () => {
@@ -121,7 +120,7 @@ describe('lineIntersection', () => {
             a: [-3, -1, -1],
             b: [1, -1, -1]
         };
-        expect(lineIntersection(negativeBox, line)).toBe(0.25);
+        expect(Aabb.lineIntersection(negativeBox, line)).toBe(0.25);
     });
 
     describe('with radius parameter', () => {
@@ -130,7 +129,7 @@ describe('lineIntersection', () => {
                 a: [-1, 1, 1],
                 b: [3, 1, 1]
             };
-            expect(lineIntersection(box, line)).toBe(lineIntersection(box, line, 0));
+            expect(Aabb.lineIntersection(box, line)).toBe(Aabb.lineIntersection(box, line, 0));
         });
 
         it('should intersect when line is outside but within radius', () => {
@@ -139,10 +138,10 @@ describe('lineIntersection', () => {
                 b: [3, 1, 1]
             };
             // Without radius, this line intersects at alpha = 0.25
-            expect(lineIntersection(box, line)).toBe(0.25);
+            expect(Aabb.lineIntersection(box, line)).toBe(0.25);
             // With radius 0.5, expanded box is [-0.5, -0.5, -0.5] to [2.5, 2.5, 2.5]
             // Line from -1 to 3 intersects at alpha = 0.125 (from -1 to -0.5 is 0.5 units, total line is 4 units)
-            expect(lineIntersection(box, line, 0.5)).toBe(0.125);
+            expect(Aabb.lineIntersection(box, line, 0.5)).toBe(0.125);
         });
 
         it('should intersect when line passes near corner with sufficient radius', () => {
@@ -151,10 +150,10 @@ describe('lineIntersection', () => {
                 b: [3, 3, 1]
             };
             // Without radius, this line intersects at alpha = 0.25
-            expect(lineIntersection(box, line)).toBe(0.25);
+            expect(Aabb.lineIntersection(box, line)).toBe(0.25);
             // With radius 0.5, expanded box is [-0.5, -0.5, -0.5] to [2.5, 2.5, 2.5]
             // Line from (-1, -1, 1) to (3, 3, 1) intersects at alpha = 0.125
-            expect(lineIntersection(box, line, 0.5)).toBe(0.125);
+            expect(Aabb.lineIntersection(box, line, 0.5)).toBe(0.125);
         });
 
         it('should still not intersect when line is too far even with radius', () => {
@@ -164,7 +163,7 @@ describe('lineIntersection', () => {
             };
             // Even with radius 0.5, expanded box is [-0.5, -0.5, -0.5] to [2.5, 2.5, 2.5]
             // Line from -2 to 4 should intersect at alpha = 0.25
-            expect(lineIntersection(box, line, 0.5)).toBe(0.25);
+            expect(Aabb.lineIntersection(box, line, 0.5)).toBe(0.25);
         });
 
         it('should handle radius larger than box size', () => {
@@ -174,7 +173,7 @@ describe('lineIntersection', () => {
             };
             // With radius 3, the expanded box is [-3, -3, -3] to [5, 5, 5]
             // Line from -5 to 7 should intersect at alpha = 1/6 (from -5 to -3 is 2 units, total line is 12 units)
-            expect(lineIntersection(box, line, 3)).toBe(1/6);
+            expect(Aabb.lineIntersection(box, line, 3)).toBe(1 / 6);
         });
 
         it('should work with negative radius (shrinks box)', () => {
@@ -183,10 +182,10 @@ describe('lineIntersection', () => {
                 b: [1.5, 1, 1]
             };
             // Without radius, this line intersects
-            expect(lineIntersection(box, line)).toBe(0);
+            expect(Aabb.lineIntersection(box, line)).toBe(0);
             // With negative radius -0.5, the box shrinks to [0.5, 0.5, 0.5] to [1.5, 1.5, 1.5]
             // Line from 0.5 to 1.5 should still intersect
-            expect(lineIntersection(box, line, -0.5)).toBe(0);
+            expect(Aabb.lineIntersection(box, line, -0.5)).toBe(0);
         });
 
         it('should handle radius with line parallel to each axis', () => {
@@ -195,21 +194,21 @@ describe('lineIntersection', () => {
                 a: [-0.5, 1, 1],
                 b: [2.5, 1, 1]
             };
-            expect(lineIntersection(box, lineX, 0.5)).toBe(0);
+            expect(Aabb.lineIntersection(box, lineX, 0.5)).toBe(0);
 
             // Y-axis
             const lineY: Line3 = {
                 a: [1, -0.5, 1],
                 b: [1, 2.5, 1]
             };
-            expect(lineIntersection(box, lineY, 0.5)).toBe(0);
+            expect(Aabb.lineIntersection(box, lineY, 0.5)).toBe(0);
 
             // Z-axis
             const lineZ: Line3 = {
                 a: [1, 1, -0.5],
                 b: [1, 1, 2.5]
             };
-            expect(lineIntersection(box, lineZ, 0.5)).toBe(0);
+            expect(Aabb.lineIntersection(box, lineZ, 0.5)).toBe(0);
         });
 
         it('should handle radius with line that starts inside expanded box', () => {
@@ -219,7 +218,7 @@ describe('lineIntersection', () => {
             };
             // Line starts at -0.5, but with radius 0.5, expanded box starts at -0.5
             // So line starts exactly at the boundary of expanded box
-            expect(lineIntersection(box, line, 0.5)).toBe(0);
+            expect(Aabb.lineIntersection(box, line, 0.5)).toBe(0);
         });
 
         it('should handle radius with line that ends inside expanded box', () => {
@@ -230,7 +229,7 @@ describe('lineIntersection', () => {
             // Line ends at 2.5, but with radius 0.5, expanded box ends at 2.5
             // Line from -1 to 2.5 (3.5 units total), intersection at -0.5 (0.5 units from start)
             // Alpha = 0.5 / 3.5 = 1/7 ≈ 0.142857
-            expect(lineIntersection(box, line, 0.5)).toBe(1/7);
+            expect(Aabb.lineIntersection(box, line, 0.5)).toBe(1 / 7);
         });
     });
 });
