--- src/math/aabb/line-intersection.ts
@@ -1,76 +1,0 @@
-/*MIT License
-
-© Copyright 2025 Adobe. All rights reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.*/
-
-
-import { Line3 } from "../line3/line3.js";
-import { Aabb } from "./aabb.js";
-
-/**
- * @returns the alpha distance along the line (0-1) where the line intersects the box or -1 if it does not.
- * @param radius - The radius of the line (treats line as a cylinder). Defaults to 0.
- */
-export function lineIntersection(box: Aabb, line: Line3, radius: number = 0): number {
-    const { min, max } = box;
-    const { a, b } = line;
-    
-    // Expand the box by the radius to treat the line as a cylinder
-    const expandedMin = [min[0] - radius, min[1] - radius, min[2] - radius];
-    const expandedMax = [max[0] + radius, max[1] + radius, max[2] + radius];
-    
-    // Calculate line direction vector
-    const dir = [b[0] - a[0], b[1] - a[1], b[2] - a[2]];
-    
-    // Calculate intersection parameters for each axis
-    const tMinX = (expandedMin[0] - a[0]) / dir[0];
-    const tMaxX = (expandedMax[0] - a[0]) / dir[0];
-    const tMinY = (expandedMin[1] - a[1]) / dir[1];
-    const tMaxY = (expandedMax[1] - a[1]) / dir[1];
-    const tMinZ = (expandedMin[2] - a[2]) / dir[2];
-    const tMaxZ = (expandedMax[2] - a[2]) / dir[2];
-    
-    // Handle division by zero (line parallel to axis)
-    const tMinXFinal = dir[0] >= 0 ? tMinX : tMaxX;
-    const tMaxXFinal = dir[0] >= 0 ? tMaxX : tMinX;
-    const tMinYFinal = dir[1] >= 0 ? tMinY : tMaxY;
-    const tMaxYFinal = dir[1] >= 0 ? tMaxY : tMinY;
-    const tMinZFinal = dir[2] >= 0 ? tMinZ : tMaxZ;
-    const tMaxZFinal = dir[2] >= 0 ? tMaxZ : tMinZ;
-    
-    // Find the largest minimum and smallest maximum
-    const tMin = Math.max(tMinXFinal, tMinYFinal, tMinZFinal);
-    const tMax = Math.min(tMaxXFinal, tMaxYFinal, tMaxZFinal);
-    
-    // Check if intersection exists
-    if (tMax < tMin || tMax < 0) {
-        return -1;
-    }
-    
-    // If tMin is negative, the line starts inside the box
-    const alpha = tMin < 0 ? 0 : tMin;
-    
-    // Check if the intersection point is within the line segment (0-1)
-    if (alpha > 1) {
-        return -1;
-    }
-    
-    return alpha;
-}
