+++ src/math/line2/intersects.test.ts
@@ -22,10 +22,9 @@ SOFTWARE.*/
 
 
 import { describe, it, expect } from 'vitest';
-import { intersects } from './intersects.js';
-import { Line2 } from './line2.js';
+import { Line2 } from '../index.js';
 
-describe('intersects', () => {
+describe('Line2.intersects', () => {
     it('should return true for intersecting lines', () => {
         const line1: Line2 = {
             a: [0, 0],
@@ -35,7 +34,7 @@ describe('intersects', () => {
             a: [0, 2],
             b: [2, 0]
         };
-        expect(intersects(line1, line2)).toBe(true);
+        expect(Line2.intersects(line1, line2)).toBe(true);
     });
 
     it('should return false for parallel lines', () => {
@@ -47,7 +46,7 @@ describe('intersects', () => {
             a: [0, 1],
             b: [2, 3]
         };
-        expect(intersects(line1, line2)).toBe(false);
+        expect(Line2.intersects(line1, line2)).toBe(false);
     });
 
     it('should return false for non-intersecting lines', () => {
@@ -59,7 +58,7 @@ describe('intersects', () => {
             a: [3, 0],
             b: [5, 2]
         };
-        expect(intersects(line1, line2)).toBe(false);
+        expect(Line2.intersects(line1, line2)).toBe(false);
     });
 
     it('should return true for lines that intersect at endpoints', () => {
@@ -71,8 +70,8 @@ describe('intersects', () => {
             a: [2, 2],
             b: [4, 4]
         };
-        expect(intersects(line1, line2, false)).toBe(false);
-        expect(intersects(line1, line2, true)).toBe(true);
+        expect(Line2.intersects(line1, line2, false)).toBe(false);
+        expect(Line2.intersects(line1, line2, true)).toBe(true);
     });
 
     it('should return false for collinear lines that do not overlap', () => {
@@ -84,7 +83,7 @@ describe('intersects', () => {
             a: [3, 3],
             b: [5, 5]
         };
-        expect(intersects(line1, line2)).toBe(false);
+        expect(Line2.intersects(line1, line2)).toBe(false);
     });
 
     it('should handle vertical and horizontal lines', () => {
@@ -96,7 +95,7 @@ describe('intersects', () => {
             a: [0, 1],
             b: [2, 1]
         };
-        expect(intersects(vertical, horizontal)).toBe(true);
+        expect(Line2.intersects(vertical, horizontal)).toBe(true);
     });
 
     it('should handle lines with negative coordinates', () => {
@@ -108,6 +107,6 @@ describe('intersects', () => {
             a: [-2, 2],
             b: [2, -2]
         };
-        expect(intersects(line1, line2)).toBe(true);
+        expect(Line2.intersects(line1, line2)).toBe(true);
     });
 });
