+++ src/math/line3/line3.ts
@@ -20,13 +20,13 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.*/
 
-
 import { FromSchema, Schema } from "../../schema/index.js";
+import { getStructLayout } from "../../typed-buffer/index.js";
 import { Vec3 } from "../index.js";
 
 const { schema: Vec3Schema } = Vec3;
 
-export const Line3Schema = {
+export const schema = {
     type: 'object',
     properties: {
         a: Vec3Schema,
@@ -34,6 +34,59 @@ export const Line3Schema = {
     },
     required: ['a', 'b'],
     additionalProperties: false,
+    default: {
+        a: Vec3Schema.default,
+        b: Vec3Schema.default,
+    }
 } as const satisfies Schema;
+export type Type = FromSchema<typeof schema>;
+type Line3 = Type;
 
-export type Line3 = FromSchema<typeof Line3Schema>; 
+export const layout = getStructLayout(schema);
+
+// Line3 utility functions
+export const interpolate = (line: Line3, alpha: number): Vec3 => {
+    const a = line.a;
+    const b = line.b;
+    return [a[0] + alpha * (b[0] - a[0]), a[1] + alpha * (b[1] - a[1]), a[2] + alpha * (b[2] - a[2])];
+};
+
+/**
+ * Calculates the alpha value (0-1) representing the closest point on a line to a given point.
+ * 
+ * @param line - The line segment
+ * @param point - The point in 3D space
+ * @returns Alpha value, usually between 0 and 1, where 0 is line.a and 1 is line.b
+ */
+export const closestPointOnLine = (line: Line3, point: Vec3): number => {
+    const { a, b } = line;
+
+    // Calculate the direction vector of the line
+    const lineDirection = Vec3.subtract(b, a);
+
+    // Calculate the vector from line start to the point
+    const pointToStart = Vec3.subtract(point, a);
+
+    // Calculate the dot product to find the projection
+    const dotProduct = Vec3.dot(pointToStart, lineDirection);
+    const lineLengthSquared = Vec3.dot(lineDirection, lineDirection);
+
+    // Avoid division by zero
+    if (lineLengthSquared === 0) {
+        return 0;
+    }
+
+    // Calculate alpha (projection parameter)
+    const alpha = dotProduct / lineLengthSquared;
+
+    // Do not clamp alpha, it can be outside of [0, 1]
+    // User can clamp after if they want to.
+    return alpha;
+};
+
+export const subLine = (line: Line3, alpha: number, beta: number): Line3 => {
+    return {
+        a: interpolate(line, alpha),
+        b: interpolate(line, beta),
+    };
+};
